Handle YouTube URLs passed to “Tune In URL”
A number of plugins allow you to give them a webpage URL, and they’ll play the music associated with that webpage, e.g., the Band’s Campout plugin has a “Bandcamp URL” field, the YouTube plugin has a “YouTube URL or Video id” field, etc.  However, it would be more convenient if you could paste the URL into the “Tune In URL” field and have LMS work out for you which plugin to use.

This commit provides that for YouTube videos, playlists, and channels.

See https://github.com/mavit/slimserver/tree/rich-url-handler for the
support for this feature in LMS.